removed frontend related code from accounts
